requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt
azure-core==1.30.1
    # via
    #   azure-identity
    #   azure-keyvault-keys
    #   azure-keyvault-secrets
    #   azure-storage-blob
    #   azure-storage-file-datalake
azure-identity==1.15.0
azure-keyvault-keys==4.9.0
azure-keyvault-secrets==4.8.0
azure-storage-blob==12.19.1
    # via azure-storage-file-datalake
azure-storage-file-datalake==12.14.0
certifi==2024.2.2
    # via requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
colorama==0.4.6
    # via pytest
cryptography==42.0.5
    # via
    #   azure-identity
    #   azure-keyvault-keys
    #   azure-storage-blob
    #   msal
    #   pyjwt
exceptiongroup==1.2.0
    # via pytest
idna==3.6
    # via requests
iniconfig==2.0.0
    # via pytest
isodate==0.6.1
    # via
    #   azure-keyvault-keys
    #   azure-keyvault-secrets
    #   azure-storage-blob
    #   azure-storage-file-datalake
msal==1.28.0
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.1.0
    # via azure-identity
packaging==24.0
    # via
    #   msal-extensions
    #   pytest
pip==24.0
pluggy==1.4.0
    # via pytest
portalocker==2.8.2
    # via msal-extensions
py4j==0.10.9.7
    # via pyspark
pycparser==2.22
    # via cffi
pyjwt==2.8.0
    # via msal
pyspark==3.4.2
pytest==8.1.1
python-dotenv==1.0.1
pywin32==306
    # via portalocker
requests==2.31.0
    # via
    #   azure-core
    #   msal
setuptools==65.5.0
six==1.16.0
    # via
    #   azure-core
    #   isodate
tomli==2.0.1
    # via pytest
typing-extensions==4.10.0
    # via
    #   azure-core
    #   azure-keyvault-keys
    #   azure-keyvault-secrets
    #   azure-storage-blob
    #   azure-storage-file-datalake
urllib3==2.2.1
    # via requests
